refactor(frontend): decouple and extract productivity screen UI and logic
Decompose the large `productivity.tsx` widget. Extract business logic, state management, and filtering into `useProductivityViewModel` to adhere to MVVM/Clean Architecture patterns. Extract the `EventCard` component to a separate file to enforce single responsibility. Maintain behavioral integrity.
